backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
pytest==6.2.5
pytest-cov==2.12.1